+++ src/index.tsx
@@ -3,7 +3,7 @@ import { render } from "solid-js/web";
 
 import "../assets/font/pixel-art-font.css";
 import "./index.css";
-import { AppTopLevel } from "./ui/v2/App";
+import { AppTopLevel } from "./ui/App";
 
 render(
     () => (
